[Feat] pinia 코드 #236
